requirements.txt
Django==4.2
djangorestframework==3.15.2
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pillow==11.0.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
drf-yasg==1.21.8
coverage==7.6.9
flake8==7.1.1
black==24.10.0
